+++ activestorage/test/controllers/representations/redirect_controller_test.rb
@@ -42,6 +42,27 @@ class ActiveStorage::Representations::RedirectControllerWithVariantsTest < Actio
   end
 end
 
+class ActiveStorage::Blobs::NoCachingRedirectControllerWithVariantsTest < ActionDispatch::IntegrationTest
+  setup do
+    @blob = create_file_blob filename: "racecar.jpg"
+    ActiveStorage.cache_service_urls = false
+  end
+
+  teardown do
+    ActiveStorage.cache_service_urls = true
+  end
+
+  test "HTTP no caching" do
+    get rails_blob_representation_url(
+      filename: @blob.filename,
+      signed_blob_id: @blob.signed_id,
+      variation_key: ActiveStorage::Variation.encode(resize_to_limit: [100, 100]))
+
+    assert_redirected_to(/racecar\.jpg/)
+    assert_equal "no-store", response.headers["Cache-Control"]
+  end
+end
+
 class ActiveStorage::Representations::RedirectControllerWithVariantsWithStrictLoadingTest < ActionDispatch::IntegrationTest
   setup do
     @blob = create_file_blob filename: "racecar.jpg"